PUT Person fix

+++ src/main/java/home/holymiko/InvestmentScraperApp/Server/Type/DTO/create/PersonCreateDTO.java
@@ -8,7 +8,7 @@ public class PersonCreateDTO extends PersonDTO {
 
     private final Long accountId;
 
-    public PersonCreateDTO(Long id, String firstName, String middleName, String lastName, String email, int phone, Long accountId) {
+    public PersonCreateDTO(Long id, String firstName, String middleName, String lastName, String email, Long phone, Long accountId) {
         super(id, firstName, middleName, lastName, email, phone);
         this.accountId = accountId;
     }
